feat(auth): restructure token/session lifecycle
- Prepared backend for stateless client access with server-side
validation
- Introduces valkey based session store
- Adds session table to db

Frontend was given access tokens to be stored in and accessed from
global context, but only client components have access and thus server
components could not be utilised in a simple way so responsibility was
shifted solely to the backend. Frontend still recieves a jwt to allow
for potential future user context if needed.